telegram_bot_django/requirements.txt
anyio==3.6.2
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
Django==4.1.5
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
python-decouple==3.7
python-telegram-bot==20.0
requests==2.28.2
rfc3986==1.5.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
zope.interface==5.5.2
